feat: added 4 PAT Sections
